refactor: move files into folders to reduce noise.
