+++ packages/cspell-trie-lib/package.json
@@ -7,6 +7,7 @@
   "files": [
     "dist",
     "!**/__mocks__",
+    "!**/test/**",
     "!**/*.test.*",
     "!**/*.spec.*",
     "!**/*.map"
