+++ packages/cspell-trie-lib/src/lib/index.ts
@@ -2,8 +2,8 @@ export * from './trie';
 export { TrieNode, FLAG_WORD, ChildMap, TrieRoot } from './TrieNode';
 export * from './trie-util';
 export * from './walker';
-export * from './importExport';
+export { importTrie, serializeTrie, ExportOptions } from './io/importExport';
 export { buildTrie, buildTrieFast, TrieBuilder } from './TrieBuilder';
 export * from './consolidate';
-export { SuggestionResult, MaxCost, suggestionCollector, SuggestionCollector } from './suggestCollector';
+export { SuggestionResult, MaxCost, suggestionCollector, SuggestionCollector } from './suggestions/suggestCollector';
 export { parseDictionaryLines, parseDictionary } from './SimpleDictionaryParser';
