+++ packages/cspell-trie-lib/src/lib/io/importExport.test.ts
@@ -1,10 +1,10 @@
-import * as Trie from '.';
-import { serializeTrie, importTrie } from './importExport';
 import { readFile } from 'fs-extra';
-import * as path from 'path';
+import * as Trie from '..';
+import { resolveSample } from '../../test/samples';
+import { importTrie, serializeTrie } from './importExport';
 
 describe('Import/Export', () => {
-    const pSampleWords = readFile(path.join(__dirname, '..', '..', 'Samples', 'sample.txt'), 'utf8');
+    const pSampleWords = readFile(resolveSample('sample.txt'), 'utf8');
     test('tests serialize / deserialize V1', async () => {
         const sampleWords = (await pSampleWords).split('\n').filter((a) => !!a);
         const trie = Trie.createTriFromList(sampleWords);
@@ -47,7 +47,7 @@ describe('Import/Export', () => {
     });
 
     test('Unsupported version', async () => {
-        const sample = await readFile(path.join(__dirname, '..', '..', 'Samples', 'sampleV2.trie'), 'utf8');
+        const sample = await readFile(resolveSample('sampleV2.trie'), 'utf8');
         const data = sample.replace('TrieXv2', 'TrieXv9').split('\n');
         expect(() => importTrie(data)).toThrow('Unsupported version: 9');
     });
