+++ packages/cspell-trie-lib/src/lib/io/importExport.ts
@@ -1,5 +1,5 @@
 import { Sequence, genSequence } from 'gensequence';
-import { TrieRoot } from './TrieNode';
+import { TrieRoot } from '../TrieNode';
 import * as iv1 from './importExportV1';
 import * as iv2 from './importExportV2';
 import * as iv3 from './importExportV3';
