+++ packages/cspell-trie-lib/src/lib/io/importExportV3.ts
@@ -1,7 +1,7 @@
-import { TrieNode, FLAG_WORD, TrieRoot } from './TrieNode';
+import { TrieNode, FLAG_WORD, TrieRoot } from '../TrieNode';
 import { Sequence, genSequence } from 'gensequence';
-import { bufferLines } from './bufferLines';
-import { trieNodeToRoot } from './trie-util';
+import { bufferLines } from '../utils/bufferLines';
+import { trieNodeToRoot } from '../trie-util';
 
 const EOW = '$';
 const BACK = '<';
