+++ packages/cspell-trie-lib/src/lib/suggestions/orthography.ts
