+++ packages/cspell-trie-lib/src/lib/suggestions/suggest-en.test.ts
@@ -1,8 +1,8 @@
-import { readTrie } from './dictionaries.test.helper';
+import { readTrie } from '../../test/dictionaries.test.helper';
 import { genCompoundableSuggestions, suggest } from './suggest';
 import { suggestionCollector, SuggestionCollectorOptions, SuggestionResult } from './suggestCollector';
-import { createTimer } from './timer';
-import { CompoundWordsMethod } from './walker';
+import { createTimer } from '../utils/timer';
+import { CompoundWordsMethod } from '../walker';
 
 function getTrie() {
     return readTrie('@cspell/dict-en_us/cspell-ext.json');
