+++ packages/cspell-trie-lib/src/lib/suggestions/suggest.legacy.test.ts
@@ -1,9 +1,9 @@
 import * as Sug from './suggest';
 import { SuggestionResult } from './suggestCollector';
-import { Trie } from './trie';
-import { TrieNode } from './TrieNode';
-import { isWordTerminationNode } from './trie-util';
-import { walker } from './walker';
+import { Trie } from '../trie';
+import { TrieNode } from '../TrieNode';
+import { isWordTerminationNode } from '../trie-util';
+import { walker } from '../walker';
 
 describe('Validate Suggest', () => {
     test('Tests suggestions against Legacy Suggestion generator', () => {
