+++ packages/cspell-trie-lib/src/lib/suggestions/suggest.ts
@@ -1,9 +1,9 @@
-import { createSuggestionOptions, GenSuggestionOptions, SuggestionOptions } from './genSuggestionsOptions';
+import { createSuggestionOptions, GenSuggestionOptions, SuggestionOptions } from '../genSuggestionsOptions';
 import { visualLetterMaskMap } from './orthography';
 import { MaxCost, suggestionCollector, SuggestionGenerator, SuggestionResult } from './suggestCollector';
-import { TrieRoot } from './TrieNode';
-import { isWordTerminationNode } from './trie-util';
-import { CompoundWordsMethod, hintedWalker, JOIN_SEPARATOR, WORD_SEPARATOR } from './walker';
+import { TrieRoot } from '../TrieNode';
+import { isWordTerminationNode } from '../trie-util';
+import { CompoundWordsMethod, hintedWalker, JOIN_SEPARATOR, WORD_SEPARATOR } from '../walker';
 
 const baseCost = 100;
 const swapCost = 75;
