+++ packages/cspell-trie-lib/src/lib/suggestions/suggestAStar.ts
@@ -1,9 +1,9 @@
-import { TrieRoot, TrieNode } from './TrieNode';
-import { CompoundWordsMethod, JOIN_SEPARATOR, WORD_SEPARATOR } from './walker';
+import { TrieRoot, TrieNode } from '../TrieNode';
+import { CompoundWordsMethod, JOIN_SEPARATOR, WORD_SEPARATOR } from '../walker';
 import { SuggestionGenerator, suggestionCollector, SuggestionResult } from './suggestCollector';
-import { PairingHeap } from './PairingHeap';
+import { PairingHeap } from '../utils/PairingHeap';
 import { visualLetterMaskMap } from './orthography';
-import { createSuggestionOptions, GenSuggestionOptionsStrict, SuggestionOptions } from './genSuggestionsOptions';
+import { createSuggestionOptions, GenSuggestionOptionsStrict, SuggestionOptions } from '../genSuggestionsOptions';
 
 export function* genCompoundableSuggestions(
     root: TrieRoot,
