+++ packages/cspell-trie-lib/src/lib/suggestions/suggestCollector.test.ts
