+++ packages/cspell-trie-lib/src/lib/suggestions/suggestCollector.ts
@@ -1,5 +1,5 @@
-import { createTimer } from './timer';
-import { JOIN_SEPARATOR, WORD_SEPARATOR } from './walker';
+import { createTimer } from '../utils/timer';
+import { JOIN_SEPARATOR, WORD_SEPARATOR } from '../walker';
 
 const defaultMaxNumberSuggestions = 10;
