+++ packages/cspell-trie-lib/src/lib/trie.test.ts
@@ -1,7 +1,7 @@
 import { SuggestionOptions } from './genSuggestionsOptions';
 import { CompoundWordsMethod, suggestionCollector } from './index';
 import { parseDictionary } from './SimpleDictionaryParser';
-import { SuggestionCollectorOptions } from './suggestCollector';
+import { SuggestionCollectorOptions } from './suggestions/suggestCollector';
 import { defaultTrieOptions, Trie } from './trie';
 import { isWordTerminationNode, normalizeWordToLowercase, orderTrie } from './trie-util';
