+++ packages/cspell-trie-lib/src/lib/trie.ts
@@ -11,8 +11,8 @@ import {
     PartialFindOptions,
 } from './find';
 import { SuggestionOptions } from './genSuggestionsOptions';
-import { genSuggestions, suggest } from './suggest';
-import { SuggestionCollector, SuggestionResult } from './suggestCollector';
+import { genSuggestions, suggest } from './suggestions/suggest';
+import { SuggestionCollector, SuggestionResult } from './suggestions/suggestCollector';
 import { PartialTrieOptions, TrieNode, TrieOptions, TrieRoot } from './TrieNode';
 import {
     countWords,
