+++ packages/cspell-trie-lib/src/lib/utils/PairingHeap.ts
