+++ packages/cspell-trie-lib/src/lib/utils/bufferLines.ts
