+++ packages/cspell-trie-lib/src/lib/utils/secondChanceCache.ts
