+++ packages/cspell-trie-lib/src/lib/utils/timer.test.ts
